+++ tools/experimental/trt-engine-explorer/trex/__init__.py
@@ -29,4 +29,4 @@ from trex.interactive import *
 from trex.report_card import *
 from trex.compare_engines import *
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"