+++ tools/experimental/trt-engine-explorer/trex/engine_plan.py
@@ -1,11 +1,12 @@
 #
-# Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 1993-2022 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+# http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,179 +16,35 @@
 #
 
 """
-TensorRT Engine Exploration API
+TensorRT Engine Exploration API - EnginePlan
 """
 
 
-from collections import OrderedDict
-from curses import meta
-from typing import Dict, List, Tuple, BinaryIO
+import warnings
+from typing import List, Tuple
 from copy import deepcopy
-import numpy as np
-import json
 import pandas as pd
 import ntpath
 from .df_preprocessing import *
-from .activations import *
+from .layer import Layer, fold_no_ops
+from .parser import *
 
-
-class Layer:
-    def __init__(self, raw_dict):
-        self.raw_dict = raw_dict
-        self.name = raw_dict['Name']
-        try:
-            self.type = raw_dict['ParameterType']
-        except:
-            self.type = raw_dict['LayerType']
-        self.subtype = raw_dict['LayerType']
-        self.inputs = [Activation(tensor) for tensor in raw_dict['Inputs']]
-        self.outputs = [Activation(tensor) for tensor in raw_dict['Outputs']]
-        self.outputs_size_bytes = np.sum([i.size_bytes for i in self.outputs])
-        if self.inputs:
-            self.precision = self.inputs[0].precision
-            self.inputs_size_bytes = np.sum([i.size_bytes for i in self.inputs])
-        else:
-            self.inputs_size_bytes = 0
-            self.precision = None
-
-        self.total_io_size_bytes = self.inputs_size_bytes + self.outputs_size_bytes
-        self._parse_weights()
-        self.total_footprint_bytes = self.total_io_size_bytes + self.weights_size
-
-    def _parse_constant(const):
-        cnt = const['Count']
-        data_type = const['Type']
-        try:
-            data_size = dict(
-                {"Int8": 1, "Half": 2, "Float": 4, "Int32": 4})[data_type]
-        except KeyError:
-            # Backward compatbility.
-            data_size = dict(
-                {"Int8": 1, "FP16": 2, "FP32": 4, "Int32": 4})[data_type]
-        return cnt, data_type, cnt * data_size
-
-    def _parse_weights(self):
-        try:
-            self.weights_cnt, self.weights_type, self.weights_size = \
-                Layer._parse_constant(self.raw_dict['Weights'])
-        except KeyError:
-            self.weights_cnt = 0
-            self.weights_type = None
-            self.weights_size = 0
-        try:
-            self.bias_cnt, self.bias_type, self.bias_size = \
-                Layer._parse_constant(self.raw_dict['Bias'])
-        except KeyError:
-            self.bias_cnt = 0
-            self.bias_type = None
-            self.bias_size = 0
-
-    def tooltip(self):
-        tip = ""
-        for key, value in sorted(self.raw_dict.items()):
-            if key not in ['InputRegions', 'OutputRegions', 'Inputs', 'Outputs',
-                    'ParameterType', 'LayerName']:
-                tip += f"{key}:{value}\\n"
-        return tip
-
-    def __repr__(self):
-        rep = f"Layer({self.name})"
-        return rep
 
 class EnginePlan:
     def __init__(self,
         graph_file: str,
         profiling_file: str=None,
-        metadata_file: str=None
+        profiling_metadata_file: str=None,
+        build_metadata_file: str=None,
+        name: str=None,
     ):
-        def read_json(json_file: str) -> BinaryIO:
-            try:
-                data = json.load(json_file)
-            except:
-                raise ValueError(f"Could not load JSON file {json_file}")
-            return data
-
-        def read_metadata_file(metadata_file: str, device: int=0):
-            with open(metadata_file) as json_file:
-                metadata = read_json(json_file)
-                return metadata[device]
-            return None
-
-        def read_profiling_file(profiling_file: str) -> List[Dict[str, any]]:
-            perf = None
-            with open(profiling_file) as json_file:
-                perf = read_json(json_file)
-                # Clean data (remove records with short size)
-                perf = [rec for rec in perf if len(rec) == 4]
-                return perf
-
-        def read_graph_file(graph_file: str) -> List:
-            with open(graph_file) as json_file:
-                graph = read_json(json_file)
-                layers = graph['Layers']
-                try:
-                    bindings = graph['Bindings']
-                except KeyError:
-                    # Older TRT didn't include bindings
-                    bindings = list()
-                return layers, bindings
-
         def path_leaf(path):
             head, tail = ntpath.split(path)
             return tail or ntpath.basename(head)
 
-        def disambiguate_layer_names(raw_layers: List) -> List:
-            """If a layer name appears twice we need to disabmiguate it"""
-            names_cnt = {}
-            for raw_layer in raw_layers:
-                name = raw_layer['Name']
-                if name in names_cnt:
-                    names_cnt[name] += 1
-                    name += "_" + str(names_cnt[name])
-                    raw_layer['Name'] = name
-                else:
-                    names_cnt[name] = 1
-            return raw_layers
-
-        def fold_no_ops(layers: List) -> List:
-            """Remove layers of type No-Op"""
-
-            def activation_consumers_dict(layers) -> Dict[Activation, List[Layer]]:
-                """Return a dictionary of consumer-layers per activation tensor"""
-                consumers = OrderedDict()
-                for layer in layers:
-                    for input in layer.inputs:
-                        try:
-                            consumers[input.name].append(layer.name)
-                        except:
-                            consumers[input.name] = [layer.name]
-                return consumers
-
-            def move_input(src: Layer, dst: Layer, i: int=0):
-                dst.inputs[i] = src.inputs[i]
-
-            def fold(no_op: Layer):
-                try:
-                    successors = activation_consumers[no_op.outputs[0].name]
-                    for successor in successors:
-                        move_input(src=no_op, dst=ret[successor])
-                except KeyError:
-                    pass
-
-            ret = OrderedDict({layer.name: layer for layer in layers})
-            activation_consumers = activation_consumers_dict(layers)
-
-            for layer in layers:
-                if layer.type == 'NoOp':
-                    fold(layer)
-
-            # Remove the No-Op layers from the final list.
-            ret = [layer for layer in ret.values() if layer.type != 'NoOp']
-            return ret
-
         def create_layers(self, raw_layers):
             layers = [Layer(raw_layer) for raw_layer in raw_layers]
-            self.layers = fold_no_ops(layers)
+            self.layers = fold_no_ops(layers, self.bindings)
             self.all_layers = deepcopy(self.layers)
             self.layers = [layer for layer in self.layers if layer.type != 'Constant']
             return raw_layers
@@ -211,7 +68,7 @@ class EnginePlan:
                     }, inplace=True)
                 df = graph_df.join(perf_df)
             else:
-                print("Warning: profiling data was not provided.")
+                warnings.warn("Profiling data was not provided.")
                 df = graph_df
                 df['latency.pct_time'] = [0] * len(df)
                 df['latency.avg_time'] = [0] * len(df)
@@ -242,33 +99,8 @@ class EnginePlan:
             self.total_runtime = sum(
                 [avg_time for avg_time in self._df["latency.avg_time"]])
 
-        def get_device_properties(metadata_file) -> Tuple[Dict, int]:
-            try:
-                metadata = read_metadata_file(metadata_file)
-                GB_1 = 1024 * 1024 * 1024
-                mem_bus_bits = metadata['GLOBAL_MEMORY_BUS_WIDTH'] # bus width in bits
-                mem_clk = metadata['MEMORY_CLOCK_RATE'] # effective DDR KHz
-                DDR = 2 # bits per clock
-                BYTE = 8 # bits
-                mem_bandwidth = mem_bus_bits * mem_clk * DDR / BYTE
-
-                return {
-                    "Device Name": metadata['Name'],
-                    "Compute Capability": metadata['ComputeCapability'],
-                    "Cuda": metadata['CUDA_VERSION'],
-                    "Total Memory (GB)": f"{metadata['TotalMemory'] / GB_1:.2f}",
-                    "Memory bus width (bits)": mem_bus_bits,
-                    "Memory clock (GHz)": f"{mem_clk / 1e6: .3f}",
-                    "Peak memory bandwidth (GB/s)": f"{mem_bandwidth / 1e6: .1f}",
-                    "Compute clock (GHz)": f"{metadata['CLOCK_RATE'] / 1e6: .2f}",
-                    "SM count": metadata['MULTIPROCESSOR_COUNT'],
-                }, mem_bandwidth
-            except:
-                return {}, 0
-
-        self.name = path_leaf(graph_file)
-        raw_layers, self.bindings = read_graph_file(graph_file)
-        raw_layers = disambiguate_layer_names(raw_layers)
+        self.name = name or path_leaf(graph_file)
+        raw_layers, self.bindings = import_graph_file(graph_file)
         raw_layers = create_layers(self, raw_layers)
 
         self._df = None
@@ -280,7 +112,8 @@ class EnginePlan:
         graph_df = add_graph_summation_cols(graph_df, self.layers)
         self._df = merge_profiling_data(graph_df, self._raw_perf)
         compute_summary(self)
-        self.metadata, self.mem_bandwidth = get_device_properties(metadata_file)
+        self.device_properties = get_device_properties(profiling_metadata_file)
+        self.performance_summary = get_performance_summary(profiling_metadata_file)
         assert self._df is not None, f"Failed parsing plan file {graph_file}"
 
     def summary_dict(self):
@@ -295,7 +128,8 @@ class EnginePlan:
             "Weights": f"{self.total_weights_size / MB_1 :.1f} MB",
             "Activations": f"{self.total_act_size/ MB_1 :.1f} MB",
         }
-        d.update(self.metadata)
+        d.update(self.device_properties)
+        d.update(self.performance_summary)
         return d
 
     def print_summary(self):